requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.3
django-tastypie==0.14.2
idna==2.7
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
python-dateutil==2.7.5
python-mimeparse==1.6.0
pytz==2018.7
requests==2.20.1
six==1.11.0
urllib3==1.24.1